move check_task_progress to es/utils
